Complete plots for new experiment.
